+++ test/WebSites/RazorPagesWebSite/HandlerTestPage.cshtml
@@ -8,14 +8,14 @@
 
     public IActionResult OnGet()
     {
-        return View();
+        return Page();
     }
 
     public async Task<IActionResult> OnPostAsync()
     {
         await TaskCache.CompletedTask;
         MethodName = nameof(OnPostAsync);
-        return View();
+        return Page();
     }
 
     public async Task OnGetCustomer()
