+++ test/WebSites/RazorPagesWebSite/OnGetView.cshtml
@@ -5,7 +5,7 @@
     {
         Message = "From OnGet";
 
-        return View();
+        return Page();
     }
     public string Message { get; set; } = "Default";
 }
