+++ test/WebSites/RazorPagesWebSite/Pages/TagHelper/PostWithHandler.cshtml
@@ -4,12 +4,12 @@
 {
     public IActionResult OnPostEdit(int id)
     {
-        return View();
+        return Page();
     }
 
     public IActionResult OnPostDelete(int id)
     {
-        return View();
+        return Page();
     }
 }
